+++ src/sensor.rs
@@ -1,16 +1,14 @@
-//! Various sensor functions 
+//! Various sensor functions
 
-use super::*; 
+use super::*;
 
-impl<T, E> LPS22HB<T> 
+impl<T, E> LPS22HB<T>
 where
     T: Interface<Error = E>,
-{   
-
-
+{
     /// Read the device ID ("who am I")
-    pub fn get_device_id(&mut self) -> Result<u8, T::Error> {        
-        let mut data = [0u8;1];
+    pub fn get_device_id(&mut self) -> Result<u8, T::Error> {
+        let mut data = [0u8; 1];
         self.interface.read(Registers::WHO_AM_I.addr(), &mut data)?;
         let whoami = data[0];
         Ok(whoami)
@@ -18,8 +16,9 @@ where
 
     /// Raw sensor reading (3 bytes of pressure data and 2 bytes of temperature data)
     fn read_sensor_raw(&mut self) -> Result<(i32, i32), T::Error> {
-        let mut data = [0u8;5];
-        self.interface.read(Registers::PRESS_OUT_XL.addr(), &mut data)?;
+        let mut data = [0u8; 5];
+        self.interface
+            .read(Registers::PRESS_OUT_XL.addr(), &mut data)?;
         let p: i32 = (data[2] as i32) << 16 | (data[1] as i32) << 8 | (data[0] as i32);
         let t: i32 = (data[4] as i32) << 8 | (data[3] as i32);
         Ok((p, t))
@@ -27,84 +26,83 @@ where
 
     /// Calculated pressure reading in hPa
     pub fn read_pressure(&mut self) -> Result<f32, T::Error> {
-        let (p,_t) = self.read_sensor_raw()?;
+        let (p, _t) = self.read_sensor_raw()?;
         let pressure: f32 = (p as f32) / PRESS_SCALE;
         Ok(pressure)
     }
 
-    /// Calculated temperaure reading in degrees Celsius 
+    /// Calculated temperaure reading in degrees Celsius
     pub fn read_temperature(&mut self) -> Result<f32, T::Error> {
-        let (_p,t) = self.read_sensor_raw()?;
+        let (_p, t) = self.read_sensor_raw()?;
         let temperature: f32 = (t as f32) / TEMP_SCALE;
         Ok(temperature)
     }
 
-
-}
-
-/*
-
-/// Pressure sensor settings. Use this struct to configure the sensor.
-#[derive(Debug)]
-pub struct SensorSettings {    
-    /// Output data rate & power mode selection
-    pub sample_rate: ODR,    
-    pub auto_addr_inc: bool,
-
-}
-
-impl Default for SensorSettings {
-    fn default() -> Self {
-        SensorSettings {            
-            sample_rate: ODR::PowerDown,            
-            auto_addr_inc: true,
-        }
+    /// Read pressure offset value, 16-bit data that can be used to implement One-Point Calibration (OPC) after soldering.
+    pub fn read_pressure_offset(&mut self) -> Result<i16, T::Error> {
+        let mut data = [0u8; 2];
+        self.interface.read(Registers::RPDS_L.addr(), &mut data)?;
+        let o: i16 = (data[1] as i16) << 8 | (data[0] as i16);
+        Ok(o)
     }
-}
 
-impl SensorSettings {
-    
-    /// Returns `u8` to write to CTRL_REG1 (0x10)
-    /// # CTRL_REG1: [0][ODR2][ODR1][ODR0][EN_LPFP][LPFP_CFG][BDU][SIM]
-    /// - ODR[2:0] - Output data rate & power mode selection
-    /// - EN_LPFP - Enable low-pass filter on pressure data when Continuous mode is used
-    /// - LPFP_CFG - Low-pass configuration register
-    /// - BDU - Block data update
-    /// - SIM - SPI Serial Interface Mode selection
-    
-    pub fn ctrl_reg1(&self) -> u8 {
-        self.sample_rate.value()
+    /// Reboot. Refreshes the content of the internal registers stored in the Flash memory block.
+    /// At device power-up the content of the Flash memory block is transferred to the internal registers
+    /// related to the trimming functions to allow correct behavior of the device itself.
+    /// If for any reason the content of the trimming registers is modified,
+    /// it is sufficient to use this bit to restore the correct values.
+    /// At the end of the boot process the BOOT bit is set again to ‘0’ by hardware.
+    /// The BOOT bit takes effect after one ODR clock cycle.
+    pub fn reboot(&mut self) -> Result<(), T::Error> {
+        self.set_register_bit_flag(Registers::CTRL_REG2, Bitmasks::BOOT)
     }
-    
-    /// Returns `u8` to write to CTRL_REG2 (0x11)
-    /// # CTRL_REG2: [BOOT][FIFO_EN][STOP_ON_FTH][IF_ADD_INC][I2C_DIS][SWRESET][ONE_SHOT]    
-    /// - BOOT - Reboot memory content
-    /// - FIFO_EN - FIFO enable
-    /// - STOP_ON_FTH - Stop on FIFO watermark. Enable FIFO watermark level use
-    /// - IF_ADD_INC - Register address automatically incremented during a multiple byte access with a
-    /// serial interface (I2C or SPI)
-    /// - I2C_DIS - Disable I2C interface
-    /// - SWRESET - Software reset
-    /// - ONE_SHOT - One-shot enable    
-    pub fn ctrl_reg2(&self) -> u8 {
-        let mut result = 0_u8;
-        if self.auto_addr_inc {
-            result |= 1 << 4;
-        }
-        result        
+
+    /// Run software reset (resets the device to the power-on configuration, takes 4 usec)
+    pub fn software_reset(&mut self) -> Result<(), T::Error> {
+        self.set_register_bit_flag(Registers::CTRL_REG2, Bitmasks::SWRESET)
     }
+
+
+    /// Is reboot phase running?
+    pub fn reboot_runnung(&mut self) -> Result<bool, T::Error> {
+        self.is_register_bit_flag_high(Registers::INT_SOURCE, Bitmasks::BOOT_STATUS)
+    }
+
+    /// Has any interrupt event been generated? (self clearing)
+    pub fn interrupt_active(&mut self) -> Result<bool, T::Error> {
+        self.is_register_bit_flag_high(Registers::INT_SOURCE, Bitmasks::IA)
+    }
+
+    /// Has low differential pressure event been generated? (self clearing)
+    pub fn low_pressure_event_occurred(&mut self) -> Result<bool, T::Error> {
+        self.is_register_bit_flag_high(Registers::INT_SOURCE, Bitmasks::PL)
+    }
+
+    /// Has high differential pressure event been generated? (self clearing)
+    pub fn high_pressure_event_occurred(&mut self) -> Result<bool, T::Error> {
+        self.is_register_bit_flag_high(Registers::INT_SOURCE, Bitmasks::PH)
+    }
+
+    /// Has new pressure data overwritten the previous one?
+    pub fn pressure_data_overrun(&mut self) -> Result<bool, T::Error> {
+        self.is_register_bit_flag_high(Registers::STATUS, Bitmasks::P_OR)
+    }
+
+    /// Has new temperature data overwritten the previous one?
+    pub fn temperature_data_overrun(&mut self) -> Result<bool, T::Error> {
+        self.is_register_bit_flag_high(Registers::STATUS, Bitmasks::T_OR)
+    }
+
+    /// Is new pressure data available?
+    pub fn pressure_data_available(&mut self) -> Result<bool, T::Error> {
+        self.is_register_bit_flag_high(Registers::STATUS, Bitmasks::P_DA)
+    }
+
+    /// Is new temperature data available?
+    pub fn temperature_data_available(&mut self) -> Result<bool, T::Error> {
+        self.is_register_bit_flag_high(Registers::STATUS, Bitmasks::T_DA)
+    }
+
+
 }
 
- */
-
-
-
-/*
-#[test]
-fn sensor_init_values() {
-    let settings = SensorSettings::default();
-    //assert_eq!(settings.ctrl_reg5_xl(), 0b0011_1000); // [DEC_1][DEC_0][Zen_XL][Yen_XL][Zen_XL][0][0][0]
-    //assert_eq!(settings.ctrl_reg6_xl(), 0b0110_0000); // [ODR_XL2][ODR_XL1][ODR_XL0][FS1_XL][FS0_XL][BW_SCAL_ODR][BW_XL1][BW_XL0]
-    //assert_eq!(settings.ctrl_reg7_xl(), 0b0000_0000); // [HR][DCF1][DCF0][0][0][FDS][0][HPIS1]
-}
- */